requirements.txt
asgiref==3.6.0
Django==3.2
django-appconf==1.0.5
django-fsm==2.8.1
django-fsm-log==3.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.3
